stock-notification/requirements.txt
# 股票分析 Agent 依赖

# 核心依赖
akshare>=1.12.0          # 金融数据源（免费）
pandas>=1.3.0            # 数据处理
numpy>=1.20.0            # 数值计算
pyyaml>=5.4.0            # 配置解析
schedule>=1.1.0          # 定时任务
requests>=2.25.0         # HTTP 请求

# 可选依赖
# tushare>=1.2.0         # Tushare 数据源（需要积分）
# talib>=0.4.0           # 技术分析库（需要系统级安装）
